+++ src/main/java/com/zelusik/eatery/global/auth/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.global.security;
+package com.zelusik.eatery.global.auth;
 
 import com.zelusik.eatery.domain.member.service.MemberQueryService;
 import org.springframework.context.annotation.Bean;
