+++ src/main/java/com/zelusik/eatery/global/auth/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.global.security;
+package com.zelusik.eatery.global.auth;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zelusik.eatery.global.exception.dto.ErrorResponse;
@@ -6,42 +6,42 @@ import com.zelusik.eatery.global.exception.constant.CustomExceptionType;
 import com.zelusik.eatery.global.exception.util.ExceptionUtils;
 import com.zelusik.eatery.global.log.LogUtils;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @Slf4j
 @Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     /**
-     * 엔드포인트에 대해 접근 권한이 존재하지 않을 때 동작하는 handler.
+     * 인증이 필요한 엔드포인트에 대해 인증되지 않았을 때 동작하는 handler.
      *
-     * @param request               that resulted in an <code>AccessDeniedException</code>
-     * @param response              so that the user agent can be advised of the failure
-     * @param accessDeniedException that caused the invocation
+     * @param request       that resulted in an <code>AuthenticationException</code>
+     * @param response      so that the user agent can begin authentication
+     * @param authenticationException that caused the invocation
      * @throws IOException if an input or output exception occurred
      */
     @Override
-    public void handle(
+    public void commence(
             HttpServletRequest request,
             HttpServletResponse response,
-            AccessDeniedException accessDeniedException
+            AuthenticationException authenticationException
     ) throws IOException {
         log.warn(
-                "[{}] JwtAccessDeniedHandler.handle() ex={}",
+                "[{}] JwtAuthenticationEntryPoint.commence() ex={}",
                 LogUtils.getLogTraceId(),
-                ExceptionUtils.getExceptionStackTrace(accessDeniedException)
+                ExceptionUtils.getExceptionStackTrace(authenticationException)
         );
 
-        response.setStatus(FORBIDDEN.value());
+        response.setStatus(UNAUTHORIZED.value());
         response.setContentType(APPLICATION_JSON_VALUE);
         response.setCharacterEncoding("utf-8");
         response.getWriter().write(
