+++ src/main/java/com/zelusik/eatery/global/auth/service/JwtTokenCommandService.java
@@ -6,8 +6,8 @@ import com.zelusik.eatery.global.auth.entity.RefreshToken;
 import com.zelusik.eatery.global.auth.exception.RefreshTokenValidateException;
 import com.zelusik.eatery.global.auth.exception.TokenValidateException;
 import com.zelusik.eatery.global.auth.repository.RefreshTokenRepository;
-import com.zelusik.eatery.global.security.JwtTokenProvider;
-import com.zelusik.eatery.global.security.dto.JwtTokenInfoDto;
+import com.zelusik.eatery.global.auth.JwtTokenProvider;
+import com.zelusik.eatery.global.auth.dto.JwtTokenInfoDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
