+++ src/main/java/com/zelusik/eatery/global/auth/service/JwtTokenQueryService.java
@@ -1,7 +1,7 @@
 package com.zelusik.eatery.global.auth.service;
 
 import com.zelusik.eatery.global.auth.repository.RefreshTokenRepository;
-import com.zelusik.eatery.global.security.JwtTokenProvider;
+import com.zelusik.eatery.global.auth.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
