+++ src/main/java/com/zelusik/eatery/global/config/JpaConfig.java
@@ -1,6 +1,6 @@
 package com.zelusik.eatery.global.config;
 
-import com.zelusik.eatery.global.security.UserPrincipal;
+import com.zelusik.eatery.global.auth.UserPrincipal;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
