+++ src/test/java/com/zelusik/eatery/config/TestSecurityConfig.java
@@ -1,7 +1,7 @@
 package com.zelusik.eatery.config;
 
+import com.zelusik.eatery.global.auth.*;
 import com.zelusik.eatery.global.config.SecurityConfig;
-import com.zelusik.eatery.global.security.*;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Import;
