+++ src/test/java/com/zelusik/eatery/unit/domain/meeting_place/api/MeetingPlaceControllerTest.java
@@ -14,7 +14,7 @@ import com.zelusik.eatery.global.common.constant.EateryConstants;
 import com.zelusik.eatery.global.common.constant.FoodCategoryValue;
 import com.zelusik.eatery.global.kakao.dto.KakaoPlaceInfo;
 import com.zelusik.eatery.global.kakao.service.KakaoService;
-import com.zelusik.eatery.global.security.UserPrincipal;
+import com.zelusik.eatery.global.auth.UserPrincipal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
