+++ src/test/java/com/zelusik/eatery/unit/domain/place_menus/api/PlaceMenusControllerTest.java
@@ -15,7 +15,7 @@ import com.zelusik.eatery.domain.place_menus.service.PlaceMenusQueryService;
 import com.zelusik.eatery.global.common.constant.EateryConstants;
 import com.zelusik.eatery.global.common.constant.FoodCategoryValue;
 import com.zelusik.eatery.global.config.JpaConfig;
-import com.zelusik.eatery.global.security.UserPrincipal;
+import com.zelusik.eatery.global.auth.UserPrincipal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
