+++ src/test/java/com/zelusik/eatery/unit/domain/review/api/ReviewControllerTest.java
@@ -24,7 +24,7 @@ import com.zelusik.eatery.domain.review_image_menu_tag.dto.request.ReviewMenuTag
 import com.zelusik.eatery.global.common.constant.EateryConstants;
 import com.zelusik.eatery.global.common.constant.FoodCategoryValue;
 import com.zelusik.eatery.global.open_ai.service.OpenAIService;
-import com.zelusik.eatery.global.security.UserPrincipal;
+import com.zelusik.eatery.global.auth.UserPrincipal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
