+++ src/test/java/com/zelusik/eatery/unit/domain/terms_info/api/TermsInfoControllerTest.java
@@ -12,7 +12,7 @@ import com.zelusik.eatery.domain.terms_info.dto.request.AgreeToTermsRequest;
 import com.zelusik.eatery.domain.terms_info.service.TermsInfoCommandService;
 import com.zelusik.eatery.global.common.constant.EateryConstants;
 import com.zelusik.eatery.global.common.constant.FoodCategoryValue;
-import com.zelusik.eatery.global.security.UserPrincipal;
+import com.zelusik.eatery.global.auth.UserPrincipal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
