+++ src/test/java/com/zelusik/eatery/unit/global/auth/service/JwtTokenCommandServiceTest.java
@@ -6,8 +6,8 @@ import com.zelusik.eatery.global.auth.entity.RefreshToken;
 import com.zelusik.eatery.global.auth.exception.TokenValidateException;
 import com.zelusik.eatery.global.auth.repository.RefreshTokenRepository;
 import com.zelusik.eatery.global.auth.service.JwtTokenCommandService;
-import com.zelusik.eatery.global.security.JwtTokenProvider;
-import com.zelusik.eatery.global.security.dto.JwtTokenInfoDto;
+import com.zelusik.eatery.global.auth.JwtTokenProvider;
+import com.zelusik.eatery.global.auth.dto.JwtTokenInfoDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
